% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfARI-v15.9.10.R
\name{estimate.mfARI.CBFV.parameters.type.2}
\alias{estimate.mfARI.CBFV.parameters.type.2}
\title{Descripci贸n de la funci贸n
Estimates the MSE for specified duration of the transient CBFV signal
In this version, Ks is the value of the normalised CBFV at the end
of the transient period.
Would be better to limit the CBFV angle here?
This would affect the slope and straight line,
affecting the transient MSE...}
\usage{
estimate.mfARI.CBFV.parameters.type.2(time.instants, normalised.CBFV.signal,
  min.CBFV.sample, min.CBFV.time.instant, transient.CBFV.duration,
  steady.CBFV.duration, time.tol = min(diff(time.instants))/100, ...)
}
\value{
Returns a plot with the CBFV parameters for a specific Delta-Tau.
}
\description{
Descripci贸n de la funci贸n
Estimates the MSE for specified duration of the transient CBFV signal
In this version, Ks is the value of the normalised CBFV at the end
of the transient period.
Would be better to limit the CBFV angle here?
This would affect the slope and straight line,
affecting the transient MSE...
}

