% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfARI-v15.9.10.R
\name{get.ATARI.mfARI.linear.relationship}
\alias{get.ATARI.mfARI.linear.relationship}
\title{Descripci贸n de la funci贸n}
\usage{
get.ATARI.mfARI.linear.relationship()
}
\value{

}
\description{
Descripci贸n de la funci贸n
}

