% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfARI-v15.9.10.R
\name{get.bounds.of.mfARI.parameters.for.AT.decimal}
\alias{get.bounds.of.mfARI.parameters.for.AT.decimal}
\title{Descripci贸n de la funci贸n}
\usage{
get.bounds.of.mfARI.parameters.for.AT.decimal()
}
\value{

}
\description{
Descripci贸n de la funci贸n
}

