% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfARI-v15.9.10.R
\name{get.mfARI.parameters}
\alias{get.mfARI.parameters}
\title{Descripci贸n de la funci贸n}
\usage{
get.mfARI.parameters(time.instants, ABP.signal, CBFV.signal, sampling.time,
  time.release, baseline.initial.time = time.instants[1],
  baseline.final.time = time.release, min.ABP.max.delta.time = 5 * 0.8,
  transient.ABP.duration = 6, min.CBFV.max.delta.time = 8 * 0.8,
  min.transient.CBFV.duration = 1.5, max.transient.CBFV.duration = 10,
  steady.CBFV.duration = 6, min.Ks = 0, max.Ks = 1.022095,
  min.ABP.angle = 0, max.ABP.angle = 90, min.CBFV.angle = 0,
  max.CBFV.angle = 90, min.Phi = 0, max.Phi = 35.238932,
  keep.details = TRUE, ABP.rounding.digits = 2,
  normalised.ABP.rounding.digits = ABP.rounding.digits * 2,
  CBFV.rounding.digits = 2,
  normalised.CBFV.rounding.digits = CBFV.rounding.digits * 2,
  Ks.rounding.digits = 4, Phi.rounding.digits = 2,
  time.tol = sampling.time/100, ...)
}
\value{

}
\description{
Descripci贸n de la funci贸n
}

