% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfARI-v15.9.10.R
\name{get.plot.CBFV.parameters.search}
\alias{get.plot.CBFV.parameters.search}
\title{Descripci贸n de la funci贸n}
\usage{
get.plot.CBFV.parameters.search(time.instants, normalised.CBFV.signal,
  min.CBFV.sample, min.CBFV.time.instant, search.results,
  cbfv.palette = brewer.pal(n = 9, name = "Blues"), id = NULL, ...)
}
\value{
Returns a plot with the CBFV parameters for a specific Delta-Tau.
}
\description{
Descripci贸n de la funci贸n
}

