% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfARI-v15.9.10.R
\name{normalise.CBFV.signal}
\alias{normalise.CBFV.signal}
\title{Descripci贸n de la funci贸n}
\usage{
normalise.CBFV.signal(time.instants, CBFV.signal, sample.release, sampling.time,
  baseline.initial.time = time.instants[1],
  baseline.final.time = time.instants[sample.release],
  min.CBFV.max.delta.time = 8 * 0.8, time.tol = sampling.time/100)
}
\value{

}
\description{
Descripci贸n de la funci贸n
}

