% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfARI-v15.9.10.R
\name{search.mfARI.CBFV.parameters}
\alias{search.mfARI.CBFV.parameters}
\title{Descripci贸n de la funci贸n
Estimates the parameters and MSE for each possible duration of the
transient CBFV signal.
Here, if the signal is too short to cover the specified maximum
transient CBFV duration, this maximum is reduced (keeping the
specified duration of the steady CBFV signal).
For equal MSE values, it prefers the one with lower transient CBFV
duration.}
\usage{
search.mfARI.CBFV.parameters(time.instants, normalised.CBFV.signal,
  min.CBFV.sample, min.CBFV.time.instant, min.transient.CBFV.duration,
  max.transient.CBFV.duration, steady.CBFV.duration,
  time.tol = min(diff(time.instants))/100,
  estimation.function = estimate.mfARI.CBFV.parameters.type.2,
  keep.search.results = FALSE, add.search.plots = FALSE, id = NULL, ...)
}
\value{

}
\description{
Descripci贸n de la funci贸n
Estimates the parameters and MSE for each possible duration of the
transient CBFV signal.
Here, if the signal is too short to cover the specified maximum
transient CBFV duration, this maximum is reduced (keeping the
specified duration of the steady CBFV signal).
For equal MSE values, it prefers the one with lower transient CBFV
duration.
}

